# docs/api_docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Generative AI Foundational Platform API",
        "version": "0.1"
    },
    "paths": {
...
        "/prompt/template/save": {
            "post": {
                "tags": [
                    "Prompt Management"
                ],
                "summary": "Create Prompt Template",
                "description": "## Endpoint to Create a Prompt Template\nThis endpoint creates a prompt template and returns the template ID.\n\n***\n\n## Request Body\n\n| Field               | Type   | Description                      |\n|---------------------|--------|----------------------------------|\n| name                | str    | The name of the prompt template. |\n| prompt_template     | str    | The prompt template.             |\n\n***\n## Example Request Body\n\n    ```json\n\n    {\n        \"name\": \"CHATBOT_PROMPT\",\n        \"prompt_template\": \"Given the following information, answer the question. Context {context}. Question {question}\"\n    }\n\n    ```\n***\n## Response Body\n\n| Field               | Type   | Description                      |\n|---------------------|--------|----------------------------------|\n| id                  | str    | The ID of the created prompt template. |\n| name                | str    | The name of the created prompt template. |\n| prompt_template     | str    | The prompt template.             |\n| version             | int    | The version of the created prompt template. |\n\n***\n#### Errors\n\n- **500 Internal Server Error**: If there is an unexpected error during the creation of the prompt template.",
                "operationId": "create_prompt_template_prompt_template_save_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePromptTemplateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TemplateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/prompt/template/get": {
            "post": {
                "tags": [
                    "Prompt Management"
                ],
                "summary": "Get Prompt Template",
                "description": "## Endpoint to Get a Prompt Template\nThis endpoint gets the latest version of a prompt template by name.\n\n***\n## Request Body\n\n| Field               | Type   | Description                      |\n|---------------------|--------|----------------------------------|\n| name                | str    | The name of the prompt template. |\n\n***\n## Example Request Body\n\n    ```json\n\n    {\n        \"name\": \"CHATBOT_PROMPT\"\n    }\n\n    ```\n***\n## Response Body\n\n| Field               | Type   | Description                      |\n|---------------------|--------|----------------------------------|\n| id                  | str    | The ID of the prompt template. |\n| name                | str    | The name of the prompt template. |\n| prompt_template     | str    | The prompt template.             |\n| version             | int    | The version of the prompt template. |\n\n***\n#### Errors\n\n- **404 Not Found**: If the prompt template is not found.\n- **500 Internal Server Error**: If there is an unexpected error during the retrieval of the prompt template.",
                "operationId": "get_prompt_template_prompt_template_get_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetPromptTemplateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TemplateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/prompt/template/versions": {
            "post": {
                "tags": [
                    "Prompt Management"
                ],
                "summary": "Get All Prompt Template",
                "description": "## Endpoint to Get All Versions of a Prompt Template\nThis endpoint gets all versions of a prompt template by name.\n\n***\n## Request Body\n\n| Field               | Type   | Description                      |\n|---------------------|--------|----------------------------------|\n| name                | str    | The name of the prompt template. |\n\n***\n## Example Request Body\n\n    ```json\n\n    {\n        \"name\": \"CHATBOT_PROMPT\"\n    }\n\n    ```\n***\n\n## Response Body\n\n[\n    {\n        \"id\": <str>,\n        \"name\": <str>,\n        \"prompt_template\": <str>,\n        \"version\": <int>\n    }\n]",
                "operationId": "get_all_prompt_template_prompt_template_versions_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetPromptTemplateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/TemplateResponse"
                                    },
                                    "type": "array",
                                    "title": "Response Get All Prompt Template Prompt Template Versions Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
            "CreatePromptTemplateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "prompt_template": {
                        "type": "string",
                        "title": "Prompt Template"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "prompt_template"
                ],
                "title": "CreatePromptTemplateRequest"
            },
            "GetPromptTemplateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "GetPromptTemplateRequest"
            },
...
            "TemplateResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "prompt_template": {
                        "type": "string",
                        "title": "Prompt Template"
                    },
                    "version": {
                        "type": "integer",
                        "title": "Version"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "prompt_template",
                    "version"
                ],
                "title": "TemplateResponse"
            }
        }
    }
}